visualize kaggle data set - note: I think Kaggle data set f up their data (there is 10 mins and 20 mins gap in 15 mins time diff patient
